+++ WEB/Src/PerformanceCollector/Perf.Shared.NetStandard/Implementation/QuickPulse/QuickPulseServiceClient.cs
@@ -24,6 +24,8 @@
     {
         private readonly string instanceName;
 
+        private readonly string roleName;
+
         private readonly string streamId;
 
         private readonly string machineName;
@@ -51,6 +53,7 @@
         public QuickPulseServiceClient(
             Uri serviceUri,
             string instanceName,
+            string roleName,
             string streamId,
             string machineName,
             string version,
@@ -61,6 +64,7 @@
         {
             this.ServiceUri = serviceUri;
             this.instanceName = instanceName;
+            this.roleName = roleName;
             this.streamId = streamId;
             this.machineName = machineName;
             this.version = version;
@@ -234,6 +238,7 @@
                 InvariantVersion = MonitoringDataPoint.CurrentInvariantVersion,
                 // InstrumentationKey = instrumentationKey, // ikey is currently set in query string parameter
                 Instance = this.instanceName,
+                RoleName = this.roleName,
                 StreamId = this.streamId,
                 MachineName = this.machineName,
                 Timestamp = timestamp.UtcDateTime,
@@ -272,6 +277,7 @@
                     InvariantVersion = MonitoringDataPoint.CurrentInvariantVersion,
                     InstrumentationKey = instrumentationKey,
                     Instance = this.instanceName,
+                    RoleName = this.roleName,
                     StreamId = this.streamId,
                     MachineName = this.machineName,
                     Timestamp = sample.EndTimestamp.UtcDateTime,
@@ -377,6 +383,7 @@
                 request.Headers.TryAddWithoutValidation(QuickPulseConstants.XMsQpsInstanceNameHeaderName, this.instanceName);
                 request.Headers.TryAddWithoutValidation(QuickPulseConstants.XMsQpsStreamIdHeaderName, this.streamId);
                 request.Headers.TryAddWithoutValidation(QuickPulseConstants.XMsQpsMachineNameHeaderName, this.machineName);
+                request.Headers.TryAddWithoutValidation(QuickPulseConstants.XMsQpsRoleNameHeaderName, this.roleName);
                 request.Headers.TryAddWithoutValidation(QuickPulseConstants.XMsQpsInvariantVersionHeaderName,
                     MonitoringDataPoint.CurrentInvariantVersion.ToString(CultureInfo.InvariantCulture));
             }